[WCNCR00380305] TWT : Fix UINT Type Underflow
[Description]
Fix coverity issue where function
mltwtGetLinkCount has an unexpected uint_32 type.

MTK-Commit-Id: fb89dcbf026643e1bfa7caee491ebe245399b7c9

CR-Id: WCNCR00380305
Feature: TWT
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/9169716
Commit-Check: srv_check_service <srv_check_service@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Tested-by: mj.tsai <mj.tsai@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Test: srv_mspautosanity
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
CodeQL-Review-Label: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
ODB-Check: srv_neptune_adm <srv_neptune_adm@mediatek.com>
AutoUT-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_pf_nep_sanity
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>

+++ include/mgmt/twt.h
@@ -484,7 +484,7 @@ uint32_t mltwtSendSetupFrameAllInOne(
 	struct _TWT_PARAMS_T *prTWTParams,
 	PFN_TX_DONE_HANDLER pfTxDoneHandler);
 
-uint32_t mltwtGetLinkCount(
+uint8_t mltwtGetLinkCount(
 	struct ADAPTER *prAdapter,
 	struct BSS_INFO *prBssInfo,
 	uint8_t ucTWTFlowId);
